Moved a few constants to SDKWorkflowRule
