+++ temporal-testing-junit4/src/main/java/io/temporal/testing/SDKTestWorkflowRule.java
@@ -43,10 +43,15 @@ import org.slf4j.LoggerFactory;
 
 public class SDKTestWorkflowRule extends TestWorkflowRule {
 
+  public static final String NAMESPACE = "UnitTest";
   public static final String BINARY_CHECKSUM = "testChecksum";
   public static final String ANNOTATION_TASK_QUEUE = "WorkflowTest-testExecute[Docker]";
   public static final String UUID_REGEXP =
       "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}";
+  public static final String TEMPORAL_SERVICE_ADDRESS = System.getenv("TEMPORAL_SERVICE_ADDRESS");
+  // Only enable when USE_DOCKER_SERVICE is true
+  public static final Boolean USE_EXTERNAL_SERVICE =
+      Boolean.parseBoolean(System.getenv("USE_DOCKER_SERVICE"));
   // Enable to regenerate JsonFiles used for replay testing.
   public static final boolean REGENERATE_JSON_FILES = false;
   private static final List<ScheduledFuture<?>> DELAYED_CALLBACKS = new ArrayList<>();
@@ -90,7 +95,7 @@ public class SDKTestWorkflowRule extends TestWorkflowRule {
     if (REGENERATE_JSON_FILES) {
       GetWorkflowExecutionHistoryRequest request =
           GetWorkflowExecutionHistoryRequest.newBuilder()
-              .setNamespace(TestWorkflowRule.NAMESPACE)
+              .setNamespace(NAMESPACE)
               .setExecution(execution)
               .build();
       GetWorkflowExecutionHistoryResponse response =
@@ -113,7 +118,7 @@ public class SDKTestWorkflowRule extends TestWorkflowRule {
   // TODO: Refactor testEnvironment to support testing through real service to avoid this
   // switches
   public void registerDelayedCallback(Duration delay, Runnable r) {
-    if (TestWorkflowRule.USE_EXTERNAL_SERVICE) {
+    if (USE_EXTERNAL_SERVICE) {
       ScheduledFuture<?> result =
           SCHEDULED_EXECUTOR.schedule(r, delay.toMillis(), TimeUnit.MILLISECONDS);
       DELAYED_CALLBACKS.add(result);
@@ -139,7 +144,7 @@ public class SDKTestWorkflowRule extends TestWorkflowRule {
   }
 
   public void sleep(Duration d) {
-    if (SDKTestWorkflowRule.USE_EXTERNAL_SERVICE) {
+    if (USE_EXTERNAL_SERVICE) {
       try {
         Thread.sleep(d.toMillis());
       } catch (InterruptedException e) {
